Add new tests/generic/536: intermittent I/O errors must not corrupt a filesystem
Based on tests/generic/347.
In our lab we've found that if multiple iSCSI connection errors are
detected (without completely loosing the iSCSI connection) then the GFS2
filesystem becomes corrupt.
Add a test that explicitly checks for this by simulating I/O errors
deterministically with dm-thin.

Tim Smith has spotted that the root cause of the corruption in GFS2
seems to be a difference in the default GFS2 block size (4k)
and iSCSI block size (512 bytes) which breaks the requirements stated in fsck.gfs2.

Although it is expected that a filesystem might report errors and
unmount itself when it detects I/O errors, when those I/O errors are
fixed and the filesystem is remounted it should not fail with a fatal
assertion again.

CC: Mark Syms <Mark.Syms@citrix.com>
CC: Tim Smith <Tim.Smith@citrix.com>
CC: Ross Lagerwall <Ross.Lagerwall@citrix.com>
Signed-off-by: Edwin Török <edvin.torok@citrix.com>